migrator work (#8)
* mirgator poc

* more stuff

* continue

* new design

* new contracts

* comments

* style

* fixes

* organize

* testing

* tests

* multicall + etching

* remove create3

* tests, fixes, snapshots

* transfer to posm first

* remove planner

* start factory test

* PR comments

* pack struct

* auction var in tests

* pr comments

* mine a salt for factory test

* one sided position

* remove params

* more error handling

* save gas by one call to posm

* adjustments

* clean up

* organize

* test org

* remove unused errors

* subscriber, calculate sqrt price, and errors params

* new changes

* find liquidity first

* fiz fuzz test

* Planner for organization

* wrap to weth

* set price

* action constants

* organize tests

* move factory into launcher

* remove factory submodule

* remove

* make tests pass

* foundryup

* infura api key

* remove unnecessary imports

* currency transfer, internal validate function

* bytes for set price + notifier

* onNotify

* remove plan and hardcode array size

* comment + gas

* gas savings + test restructure

* make variables names clear

* test restructure

* clean up tests

+++ src/types/Distribution.sol
@@ -1,9 +1,10 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.26;
 
+/// @title Distribution
 /// @notice Represents one distribution instruction: which strategy to use, how many tokens, and any custom data
 struct Distribution {
     address strategy;
-    uint256 amount;
+    uint128 amount;
     bytes configData;
 }
